Fix for sifnoded genaccounts cmd flag parsing

+++ cmd/sifnoded/cmd/genaccounts.go
@@ -77,11 +77,11 @@ contain valid denominations. Accounts may optionally be supplied with vesting pa
 				return fmt.Errorf("failed to parse coins: %w", err)
 			}
 
-			vestingStart, err := cmd.Flags().GetInt64(flagVestingStart)
+			vestingStart, err := cmd.Flags().GetUint64(flagVestingStart)
 			if err != nil {
 				return err
 			}
-			vestingEnd, err := cmd.Flags().GetInt64(flagVestingEnd)
+			vestingEnd, err := cmd.Flags().GetUint64(flagVestingEnd)
 			if err != nil {
 				return err
 			}
@@ -101,11 +101,11 @@ contain valid denominations. Accounts may optionally be supplied with vesting pa
 			balances := banktypes.Balance{Address: addr.String(), Coins: coins.Sort()}
 			baseAccount := authtypes.NewBaseAccount(addr, nil, 0, 0)
 			if !vestingAmt.IsZero() {
-				baseVestingAccount := authvesting.NewBaseVestingAccount(baseAccount, vestingAmt.Sort(), vestingEnd)
+				baseVestingAccount := authvesting.NewBaseVestingAccount(baseAccount, vestingAmt.Sort(), int64(vestingEnd))
 
 				switch {
 				case vestingStart != 0 && vestingEnd != 0:
-					genAccount = authvesting.NewContinuousVestingAccountRaw(baseVestingAccount, vestingStart)
+					genAccount = authvesting.NewContinuousVestingAccountRaw(baseVestingAccount, int64(vestingStart))
 
 				case vestingEnd != 0:
 					genAccount = authvesting.NewDelayedVestingAccountRaw(baseVestingAccount)
